pyproject.toml
[tool.poetry]
name = "fastflows"
version = "0.1.0.dev.1656871708"
description = "FastFlows is a FastAPI server & command line tool to comunicate with Prefect 2.0 as a Workflow manager to deploy, run, track flows and more."
authors = ["Francesco Bartoli <francesco.bartoli@geobeyond.it>", "Iuliia Volkova <xnuinside@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/geobeyond/fastflows"
repository = "https://github.com/geobeyond/fastflows"
documentation = "https://github.com/geobeyond/fastflows"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.4.1"
httpx = "^0.23.0"
loguru = "^0.6.0"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
mangum = "^0.15.0"
fastapi-opa = "^1.3.7"
uvicorn = "^0.17.6"
rich = "^12.4.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^22.3.0"
nox = "^2022.1.7"
nox-poetry = "^1.0.1"

[tool.poetry.scripts]
fastflows = "fastflows.cli:app"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
